+++ app/scripts/components/advanced_search.js
@@ -34,8 +34,15 @@ export const advancedSearchComponent = {
             on-search="$ctrl.onSearch()"
             on-search-save="$ctrl.onSearchSave(name)"
         ></search-submit>
-        <input id="inOrderChkAdv" type="checkbox" ng-model="$ctrl.inOrder" ng-disabled="!$ctrl.inOrderEnabled" />
-        <label for="inOrderChkAdv"> {{'in_order_chk' | loc:$root.lang}}</label>
+        <input id="inOrderChkAdv" type="checkbox" ng-model="$ctrl.inOrder" />
+        <label for="inOrderChkAdv">
+            {{'in_order_chk' | loc:$root.lang}}
+            <i
+                class="fa fa-info-circle text-gray-400"
+                uib-tooltip="{{'order_help' | loc:$root.lang}}"
+                tooltip-placement="right"
+            ></i>
+        </label>
     </div>`,
     bindings: {},
     controller: [
@@ -46,8 +53,6 @@ export const advancedSearchComponent = {
             const $ctrl = this
 
             $ctrl.inOrder = $location.search().in_order == null
-            /** Whether the "in order" option is applicable. */
-            $ctrl.inOrderEnabled = true
 
             if ($location.search().search && $location.search().search.split("|")) {
                 var [type, ...expr] = $location.search().search.split("|")
@@ -61,14 +66,10 @@ export const advancedSearchComponent = {
             }
 
             $ctrl.onSearch = () => {
-                // The "in order" option should be available only if >1 token and no wildcards.
-                const cqpObjs = CQP.parse($ctrl.cqp)
-                $ctrl.inOrderEnabled = cqpObjs.length > 1 && !CQP.hasWildcards(cqpObjs)
-
                 $location.search("search", null)
                 $location.search("page", null)
                 $location.search("within", null)
-                $location.search("in_order", !$ctrl.inOrder && $ctrl.inOrderEnabled ? false : null)
+                $location.search("in_order", !$ctrl.inOrder ? false : null)
                 $timeout(() => $location.search("search", `cqp|${$ctrl.cqp}`), 0)
             }